replace EvaluationResult.kt with Finding simplifying the design
