+++ codyze-backends/cpg/src/main/kotlin/de/fraunhofer/aisec/codyze/backends/cpg/coko/evaluators/FollowsEvaluator.kt
@@ -16,11 +16,9 @@
 package de.fraunhofer.aisec.codyze.backends.cpg.coko.evaluators
 
 import de.fraunhofer.aisec.codyze.backends.cpg.coko.CokoCpgBackend
-import de.fraunhofer.aisec.codyze.backends.cpg.coko.CpgEvaluationResult
 import de.fraunhofer.aisec.codyze.backends.cpg.coko.CpgFinding
 import de.fraunhofer.aisec.codyze.backends.cpg.coko.dsl.getNodes
 import de.fraunhofer.aisec.codyze.specificationLanguages.coko.core.EvaluationContext
-import de.fraunhofer.aisec.codyze.specificationLanguages.coko.core.EvaluationResult
 import de.fraunhofer.aisec.codyze.specificationLanguages.coko.core.Evaluator
 import de.fraunhofer.aisec.codyze.specificationLanguages.coko.core.dsl.Op
 import de.fraunhofer.aisec.cpg.query.executionPath
@@ -28,12 +26,12 @@ import de.fraunhofer.aisec.cpg.query.executionPath
 context(CokoCpgBackend)
 @Suppress("UnusedPrivateMember") // TODO: remove once this returns actual findings
 class FollowsEvaluator(val ifOp: Op, val thenOp: Op) : Evaluator {
-    override fun evaluate(context: EvaluationContext): EvaluationResult<CpgFinding> {
+    override fun evaluate(context: EvaluationContext): List<CpgFinding> {
         val evaluator = {
             val thisNodes = with(this@CokoCpgBackend) { ifOp.getNodes() }
             val thatNodes = with(this@CokoCpgBackend) { thenOp.getNodes() }
             thisNodes.all { from -> thatNodes.any { to -> executionPath(from, to).value } }
         }
-        return CpgEvaluationResult(listOf(CpgFinding("TODO")))
+        return listOf(CpgFinding("TODO"))
     }
 }
